requirements.txt
fastapi
uvicorn[standard]
librosa
soundfile
python-multipart
aiofiles
torch
torchaudio
transformers
pyannote.audio
panns-inference
whisper
pandas
numpy
